refactor: model層を別ディレクトリに切り分け
