+++ app/api/detail/[contentId]/routeClient.ts
@@ -1,6 +1,6 @@
 import { filterPublicReplies } from '@/lib'
 import { QA } from '@/types'
-import { fetchPost } from '../../microcms'
+import { fetchPost } from '../../../models/microcms'
 import { noResultQAData } from '@/consts'
 
 export async function GET_CLIENT(contentId: string): Promise<QA> {
